requirements.txt
aniso8601==7.0.0
attrs==24.2.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-GraphQL==2.0.1
flask-restx==1.3.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
promise==2.3
psycopg2==2.9.10
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
referencing==0.35.1
req==1.0.0
requests==2.32.3
rpds-py==0.21.0
Rx==1.6.3
simple-websocket==1.1.0
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
